refactor: extract reusable parts of the client engine into a new package
`@prisma/client-engine-runtime` being built by `@prisma/client`'s build
script meant that we still needed to include the client package and all
of its transitive dependencies into the workspace in
`libs/driver-adapters` in the `prisma-engines` repo, which was causing
some issues.

This commit properly extracts the reusable parts of the client engine
into a package for real. This is also important to make Accelerate work
with the query compiler.

Additionally, `@prisma/debug` is changed to be ESM compatible. This is
because `@prisma/client-engine-runtime` needs to be ESM compatible and
depends on `@prisma/debug`.